+++ Editor/Expressions/SearchExpressionImporter.cs
@@ -5,7 +5,6 @@ using UnityEngine.UIElements;
 
 #if UNITY_2020_2_OR_NEWER
 using UnityEditor.AssetImporters;
-using UnityEditor.Experimental.AssetImporters;
 #else
 using UnityEditor.Experimental.AssetImporters;
 #endif