+++ Editor/Indexing/SearchDatabaseImporter.cs
@@ -4,7 +4,6 @@ using UnityEngine;
 
 #if UNITY_2020_2_OR_NEWER
 using UnityEditor.AssetImporters;
-using UnityEditor.Experimental.AssetImporters;
 #else
 using UnityEditor.Experimental.AssetImporters;
 #endif